refactor.

+++ Crystal/AppBase/PSBoxView.cpp
@@ -28,7 +28,7 @@ void PSBoxView::onOk()
 	auto scene = new ParticleSystemScene();
 
 	std::mt19937 mt{ std::random_device{}() };
-	std::uniform_real_distribution<double> dist(0.0, 1.0);
+	std::uniform_real_distribution<float> dist(0.0, 1.0);
 	for (int i = 0; i < countView.getValue(); ++i) {
 		const auto u = dist(mt);
 		const auto v = dist(mt);
